Break into separate flies
